Move SubmissionFSM Events into Persistence-Model
This will allow us to make protobuf serializers for these events
