+++ persistence/src/main/scala/hmda/persistence/processing/SubmissionFSM.scala
@@ -11,7 +11,8 @@ import hmda.model.fi.{ Submission, SubmissionId, SubmissionStatus }
 import hmda.persistence.HmdaSupervisor.FindSubmissions
 import hmda.persistence.institutions.SubmissionPersistence
 import hmda.persistence.institutions.SubmissionPersistence.UpdateSubmissionStatus
-import hmda.persistence.messages.CommonMessages.{ Command, Event, GetState }
+import hmda.persistence.messages.CommonMessages.{ Command, GetState }
+import hmda.persistence.messages.events.processing.SubmissionFSMEvents._
 import hmda.persistence.processing.ProcessingMessages.{ CompleteValidationWithErrors, Sign, _ }
 import hmda.persistence.processing.SubmissionFSM.{ Signed, _ }
 
@@ -25,26 +26,8 @@ object SubmissionFSM {
 
   val failedMsg = "Submission status update failed"
 
-  trait SubmissionEvent extends Event
-
-  trait SubmissionState
-
   //Commands
   case object Create extends Command
-
-  //Domain Events (persisted)
-  case class SubmissionCreated(s: Submission) extends SubmissionEvent
-  case class SubmissionUploading(s: Submission) extends SubmissionEvent
-  case class SubmissionUploaded(s: Submission) extends SubmissionEvent
-
-  case class SubmissionParsing(s: Submission) extends SubmissionEvent
-  case class SubmissionParsed(s: Submission) extends SubmissionEvent
-  case class SubmissionParsedWithErrors(s: Submission) extends SubmissionEvent
-  case class SubmissionValidating(s: Submission) extends SubmissionEvent
-  case class SubmissionValidated(s: Submission) extends SubmissionEvent
-  case class SubmissionValidatedWithErrors(s: Submission) extends SubmissionEvent
-  case class SubmissionSigned(s: Submission) extends SubmissionEvent
-  case class SubmissionFailed(s: Submission) extends SubmissionEvent
 
   //Submission States
   sealed trait SubmissionFSMState extends FSMState
